add extern sources support
if project uses packages, which sources are out of $GOROOT/src/pkg, they must be listed in '.externPackages' section of goref.cfg. Example can be seen in testPack. All packages, recursively used by extern packages, must be listed either.
Column 2 defines 'goPath'. This is the path, which is used in import declaration. It can point at a package in $GOROOT/pkg, or to any other package (in this case perhaps there is special build script for application)